add SciMLStructures and relax params type
